+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The `Unreleased` section name is replaced by the expected version of next releas
 ### Removed
 ### Fixed
 
+- `UnionConverter`: Handle nested unions [#52](https://github.com/jet/FsCodec/pull/52)
 - `UnionConverter`: Support overriding discriminator without needing to nominate a `catchAllCase` [#51](https://github.com/jet/FsCodec/pull/51)
 
 <a name="2.1.0"></a>